requirements.txt
# Web界面
streamlit>=1.28.0

# 文档处理
python-docx>=0.8.11

# 数据处理
pandas>=1.5.0
numpy>=1.21.0

# 自然语言处理
nltk>=3.8.0

# 机器学习和文本相似度
scikit-learn>=1.3.0

# 配置管理
pyyaml>=6.0

# HTTP请求
requests>=2.31.0

# 可选：用于增强的文本分析
# spacy>=3.7.0

# ========== 混合检索引擎依赖（方案4）==========
# 高性能向量检索（推荐安装，显著提升性能）
faiss-cpu>=1.7.4
# 如使用NVIDIA GPU，可改用: faiss-gpu>=1.7.4

# 向量嵌入和Cross-encoder重排序
sentence-transformers>=2.2.0

# 注意：首次使用时会自动下载以下模型（约100-200MB）
# - sentence-transformers/all-MiniLM-L6-v2 (向量编码)
# - cross-encoder/ms-marco-MiniLM-L-6-v2 (重排序)
